remove meroxa-py dep
