+++ setup.cfg
@@ -26,7 +26,6 @@ python_requires = >=3.7
 install_requires =
     aiohttp>=3.8
     grpcio>=1.44.0
-    meroxa-py
 
 [options.entry_points]
 console_scripts =
