+++ src/turbine/runner/__init__.py
@@ -1,5 +1,4 @@
 from .app_generate import generate_app
-from .baserunner import BaseRunner
 from .runner import Runner
 
-__all__ = ["generate_app", BaseRunner, Runner]
+__all__ = ["generate_app", "Runner"]
